Extended WalletUtils to allow prompting for aliases.
